requirements2.txt
antlr4-python3-runtime==4.9.3
asttokens==3.0.1
backcall==0.2.0
carla==0.9.12
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.1.8
cloudpickle==3.1.2
comm==0.2.3
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
docker-pycreds==0.4.0
einops==0.8.1
executing==2.2.1
filelock==3.16.1
fonttools==4.57.0
fsspec==2025.3.0
gitdb==4.0.12
gitpython==3.1.46
gym==0.21.0
h5py==3.7.0
Hydra==2.5
idna==3.11
imageio==2.35.1
importlib-metadata==4.13.0
importlib-resources==6.4.5
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
jinja2==3.1.6
jupyter-client==8.6.3
jupyter-core==5.8.1
Keras-Applications==1.0.8
kiwisolver==1.4.7
lazy-loader==0.4
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.1
numpy==1.23.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
omegaconf==2.2.3
opencv-python==4.6.0.66
packaging==26.0
pandas==1.5.0
parso==0.8.6
pathtools==0.1.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.3.6
promise==2.3
prompt-toolkit==3.0.52
protobuf==4.25.9
psutil==7.2.2
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.19.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2026.1.post1
PyWavelets==1.4.1
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.4
scikit-image==0.21.0
scipy==1.10.1
sentry-sdk==2.57.0
setproctitle==1.3.7
shapely==2.0.6
shortuuid==1.0.13
six==1.17.0
smmap==5.0.3
stable-baselines3==1.6.2
stack-data==0.6.3
sympy==1.13.3
tifffile==2023.7.10
torch==2.2.2
torchvision==0.17.2
tornado==6.4.2
tqdm==4.66.2
traitlets==5.14.3
triton==2.2.0
typing-extensions==4.13.2
urllib3==2.2.3
wandb==0.13.4
wcwidth==0.6.0
zipp==3.20.2
